requirements.txt
asgiref==3.8.1
awscli==1.38.35
boto3==1.37.35
botocore==1.37.35
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.1.7
django-cleanup==9.0.0
django-storages==1.14.6
djangorestframework==3.16.0
docutils==0.19
filelock==3.18.0
fsspec==2025.5.0
gunicorn==23.0.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.6
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
whitenoise==6.9.0